Do not require a black hole to be updated
This allows removing black hole handling altogether by removing the
closure -> black hole update call upon entering an updatable closure.
The downside is that it is now much harder to detect whether black hole
updating works. Since it's a very simple and easily understood
mechanism, I'm choosing the not-so-safe variant of handling the matter.