Add HTML test result generator
Run test with "--html out.html" to generate a nice test report.
Unfortunately, this is incompatible to printing the results to the
console, so you'll have to choose between HTML or not.

This also modifies the test templates to use ANSI prettyprinting in the
console, but not for the HTML reports.