Fix copying GC dead address reporting; fix accidentially correct test
The coping GC reported (addresses on old heap MINUS addresses on new
heap) as being dead. This is wrong in case of a moving garbage collector
of course, since we cannot compare old and new addresses directly
without considering the forwarding indirections!

The test wasn't good for the moving case either: the copying collection
only accidentially didn't move stuff around! Once a new closure was
added that was moved to the beginning by the GC, the expectations
failed due to having hardcoded "should be alive/dead" addresses.

The new test is weaker than the old one, testing only for heap sizes
instead of actual contents though.